fix(infra api): fix issue causing ui to show error on project create
Updating the create function to take options required passing the project documents as an array. As the documents are passed in an array, the Mongoose function returns an array whereas before it was returning just the created database document. This was causing the GraphQL layer to fail as it was receiving an array rather than a project document.