Add vscode option to frontend
In a test environment this completes the addition of vscode, although
there are issues with websockets which breaks some of the functionality.

We update all the image_config.json files to include the vscode
additions, so local development and the frontend see the new notebook
type.

Since the token acquisition in stackUrlService.js is identical to
Jupyter notebooks we refactor it into a generic function.

Use of the logo should comply with the restrictions given at
https://code.visualstudio.com/brand.

Jira: ED-45
Jira: ED-46